Logger fully functioning, only issue is that times of logs are not showing up
